Refactor ServiceHighlight components to use services.json
