+++ react-app/src/stories/Service.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 
 import "./service.css";
-import { ReactComponent as IconExt } from "../img/external-link-alt-solid.svg";
+import { ReactComponent as IconExt } from "./assets/external-link-alt-solid.svg";
 
 export const Service= ({ name, description, url, external }) => (
   <div className="div--service">
